refactor(Nav): remove the interface to index file
