--- src/components/navigation/Nav/interfaces/Nav.LinkComponet.interface.ts
@@ -1,5 +1,0 @@
-import { ILink } from "./Nav.Link.interface";
-
-export interface INavLinkProps {
-  section: ILink[];
-}
