--- src/components/navigation/Nav/interfaces/Nav.MultiSections.interface.ts
@@ -1,6 +1,0 @@
-import { INavigation } from "./Nav.Navigation.interface";
-
-export interface INavMultiSectionsProps {
-  navigation: INavigation;
-  sections: string[];
-}
