--- src/components/navigation/Nav/interfaces/Nav.Navigation.interface.ts
@@ -1,6 +1,0 @@
-import { ISection } from "./Nav.Sections.interface";
-
-export interface INavigation {
-  title: string;
-  sections: { [key: string]: ISection };
-}
