--- src/components/navigation/Nav/interfaces/Nav.OneSection.interface.ts
@@ -1,6 +1,0 @@
-import { INavigation } from "./Nav.Navigation.interface";
-
-export interface INavOneSectionProps {
-  navigation: INavigation;
-  firstSection: string;
-}
