--- src/components/navigation/Nav/interfaces/Nav.Sections.interface.ts
@@ -1,6 +1,0 @@
-import { ILink } from "./Nav.Link.interface";
-
-export interface ISection {
-  name: string;
-  links: { [key: string]: ILink };
-}
