--- src/components/navigation/Nav/interfaces/Nav.interface.ts
@@ -1,7 +1,0 @@
-import { INavigation } from "./Nav.Navigation.interface";
-
-export interface INavProps {
-  navigation: INavigation;
-  logoutPath: string;
-  logoutTitle: string;
-}
